pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecg-interpretation-system"
version = "0.1.0"
description = "AI-powered ECG interpretation system for clinical alarm events"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "h5py>=3.8.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]
llm = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]
ml = [
    "einops>=0.6.0",
    "timm>=0.9.0",
    "neurokit2>=0.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
all = [
    "ecg-interpretation-system[api,llm,ml,dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "gate: gate tests for phase boundary contracts",
    "integration: integration tests spanning multiple modules",
]
